Factor before methods to spec helper
