+++ spec/features/user_can_see_orders_on_dashboard_spec.rb
@@ -2,15 +2,6 @@ require 'rails_helper'
 
 feature 'user has order list on dashboard' do
   before do
-    def create_order
-      visit '/signin_or_signup'
-      click_link_or_button("twitter-link")
-      visit '/items'
-      click_link_or_button('pencil')
-      click_link_or_button("Add to Cart")
-      visit '/cart'
-      click_link_or_button("Checkout")
-    end
     create_order
   end
