+++ spec/features/user_provides_address_on_checkout_spec.rb
@@ -1,0 +1,10 @@
+require 'rails_helper'
+
+feature 'checkout' do
+  before do
+    create_order
+  end
+  it 'requires a mailing address on checkout' do
+
+  end
+end
