+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'omniauth'
 require 'ostruct'
 
 RSpec.configure do |config|
-
+  
   config.backtrace_exclusion_patterns = []
   config.backtrace_exclusion_patterns << /.*\/gems\/.*/
     config.before(:suite) do
@@ -60,6 +60,25 @@ RSpec.configure do |config|
     def login_user!
       visit signin_or_signup_path
       click_on "twitter-link"
+    end
+
+    def fill_cart
+      visit '/signin_or_signup'
+      click_link_or_button("twitter-link")
+      visit '/items'
+      click_link_or_button('pencil')
+      click_link_or_button("Add to Cart")
+      visit '/cart'
+    end
+
+    def create_order
+      visit '/signin_or_signup'
+      click_link_or_button("twitter-link")
+      visit '/items'
+      click_link_or_button('pencil')
+      click_link_or_button("Add to Cart")
+      visit '/cart'
+      click_link_or_button("Checkout")
     end
 
   # rspec-expectations config goes here. You can use an alternate
